Revert "Move console form to ui/console/form package"
This reverts commit 6ade203222d2c3110b134ebdb96c8a1f0d8bad4a.
