+++ cmd/login_cloud.go
@@ -11,7 +11,7 @@ import (
 
 	"go.k6.io/k6/cloudapi"
 	"go.k6.io/k6/lib/consts"
-	"go.k6.io/k6/ui/console/form"
+	"go.k6.io/k6/ui"
 )
 
 //nolint:funlen,gocognit
@@ -70,13 +70,13 @@ This will set the default token used when just "k6 run -o cloud" is passed.`,
 			case token.Valid:
 				newCloudConf.Token = token
 			default:
-				f := form.Form{
-					Fields: []form.Field{
-						form.StringField{
+				form := ui.Form{
+					Fields: []ui.Field{
+						ui.StringField{
 							Key:   "Email",
 							Label: "Email",
 						},
-						form.PasswordField{
+						ui.PasswordField{
 							Key:   "Password",
 							Label: "Password",
 						},
@@ -86,7 +86,7 @@ This will set the default token used when just "k6 run -o cloud" is passed.`,
 					globalState.logger.Warn("Stdin is not a terminal, falling back to plain text input")
 				}
 				var vals map[string]string
-				vals, err = f.Run(globalState.console.Stdin, globalState.console.Stdout)
+				vals, err = form.Run(globalState.console.Stdin, globalState.console.Stdout)
 				if err != nil {
 					return err
 				}
