--- ui/console/form/doc.go
@@ -1,3 +1,0 @@
-// Package form implements elements for handling plain text and password inputs
-// in the console.
-package form
