+++ ui/form_fields.go
@@ -1,4 +1,4 @@
-package form
+package ui
 
 import (
 	"bufio"
