+++ ui/form_test.go
@@ -1,4 +1,4 @@
-package form
+package ui
 
 import (
 	"bytes"
